Renamed the utils and constants to editorUtils for easier readability.
