+++ src/components/EditorContent/utils/editorUtils.js
